Customer behind Truck Complete

+++ src/main/java/com/heycream/AbstractAndInterface/PricedItem.java
@@ -1,7 +1,5 @@
-
 package com.heycream.AbstractAndInterface;
 
-
 public interface PricedItem {
-    double getPrice();
+    int getPrice();
 }
